+++ eth_defi/confirmation.py
@@ -8,6 +8,7 @@
 import datetime
 import logging
 import time
+from _decimal import Decimal
 from typing import Dict, List, Set, Union, cast, Collection, TypeAlias
 
 from eth_account.datastructures import SignedTransaction
@@ -39,8 +40,24 @@ class ConfirmationTimedOut(Exception):
     """We exceeded the transaction confirmation timeout."""
 
 
+class NonRetryableBroadcastException(Exception):
+    """Don't try to rebroadcast these."""
+
+
 class NonceMismatch(Exception):
     """Chain has a different nonce than we expect."""
+
+
+class OutOfGasFunds(NonRetryableBroadcastException):
+    """Out of gas funds for an executor."""
+
+
+class NonceTooLow(NonRetryableBroadcastException):
+    """Out of gas funds for an executor."""
+
+
+class BadChainId(NonRetryableBroadcastException):
+    """Out of gas funds for an executor."""
 
 
 def wait_transactions_to_complete(
@@ -367,9 +384,34 @@ def _broadcast_multiple_nodes(providers: Collection[BaseProvider], signed_tx: Si
             logger.info("Source: %s", source)
 
             # When we rebroadcast we are getting nonce too low errors,
-            # both for too high and too low nonces
+            # both for too high and too low nonces.
+            # We also get nonce too low errors,
+            # when broadcasting through multiple nodes and those nodes sync nonce faster than we broadcast
             if resp_data["message"] == "nonce too low":
-                continue
+                if address:
+                    current_nonce = web3.eth.get_transaction_count(address)
+                else:
+                    current_nonce = None
+
+                logger.info("Nonce too low. Current:%s proposed:%s address:%s: tx:%s resp:%s", current_nonce, nonce, address, signed_tx, resp_data)
+                #raise NonceTooLow(f"Current on-chain nonce {current_nonce}, proposed {nonce}") from e
+
+            if "invalid chain" in resp_data["message"]:
+                # Invalid chain id / chain id missing.
+                # Cannot retry.
+                logger.warning("Invalid chain: %s %s", signed_tx, resp_data)
+                raise BadChainId() from e
+
+            if "insufficient funds for gas" in resp_data["message"]:
+                logger.warning("Out of balance error. Tx: %s, resp: %s", signed_tx, resp_data)
+                # Always raise when we are out of funds,
+                # because any retry is not help
+                if address:
+                    our_balance = web3.eth.get_balance(address)
+                    our_balance = Decimal(our_balance) / Decimal(10**18)
+                else:
+                    our_balance = None
+                raise OutOfGasFunds(f"Failed to broadcast {tx_hash}, out of gas, account {address} balance is {our_balance}.\n" f"TX details: {signed_tx}") from e
 
         except Exception as e:
             exceptions[p] = e
@@ -484,6 +526,10 @@ def wait_and_broadcast_multiple_nodes(
 
         In the case of multiple exceptions, the last one is raised.
         The exception is whatever lower stack is giving us.
+
+    :raise OutOfGasFunds:
+        The hot wallet account does not have enough native token to cover the tx fees.
+
     """
 
     assert isinstance(poll_delay, datetime.timedelta)
@@ -538,6 +584,9 @@ def wait_and_broadcast_multiple_nodes(
         try:
             _broadcast_multiple_nodes(providers, tx)
             last_exception = None
+        except NonRetryableBroadcastException:
+            # Don't try to handle
+            raise
         except Exception as e:
             last_exception = e
 
@@ -580,7 +629,6 @@ def wait_and_broadcast_multiple_nodes(
         unconfirmed_txs -= confirmation_received
 
         if unconfirmed_txs:
-
             # TODO: Clean this up after the root cause with Anvil is figured out
             if mine_blocks:
                 timestamp = get_latest_block_timestamp(web3)
@@ -611,9 +659,7 @@ def wait_and_broadcast_multiple_nodes(
                         logger.exception(e)
 
                 unconfirmed_tx_strs = ", ".join([tx_hash.hex() for tx_hash in unconfirmed_txs])
-                raise ConfirmationTimedOut(
-                    f"Transaction confirmation failed. Started: {started_at}, timed out after {max_timeout} ({max_timeout.total_seconds()}s). Poll delay: {poll_delay.total_seconds()}s. Still unconfirmed: {unconfirmed_tx_strs}"
-                )
+                raise ConfirmationTimedOut(f"Transaction confirmation failed. Started: {started_at}, timed out after {max_timeout} ({max_timeout.total_seconds()}s). Poll delay: {poll_delay.total_seconds()}s. Still unconfirmed: {unconfirmed_tx_strs}")
 
         if datetime.datetime.utcnow() >= next_node_switch:
             # Check if it time to try a better node provider