+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Current
 
+- Fix: `HotWallet.sign_bound_call_with_new_nonce` tries to avoid calling broken Web3 gas estimation
+  machine if the gas parameters are already given as the arguments
+- Fix: Raise `OutOfGasFunds` in `_broadcast_multiple_nodes` and
+  avoid transaction broadcast retry if we do not have gas money
+- Fix: Don't swallow nonce errors and chain id errors in `broadcast_multiple_nodes`
 - Fix type normalisation of `tx_hash` in `fetch_transaction_revert_reason`
 
 # 0.24.4